Added ability to run executable on mutated inputs
run-on-fuzzed-input allows easy way to run build on mutated input for a series
of bugs.

This uses signal_analysis_gdb.py which is a gdb python scripts that is mean to
be given to gdb through -x option.  This gdb script handles all signals
and prints info proc mappings and the memory violation address when a SIGSEGV
happens