Added checks for installed packages and dock perms
setup.py is now doesn't require sudo after the first run.
this is convenient in that if you already have your lava environment
set up and working, you may want to check that it is consistent
by running `python setup.py`.  If sudo request is asked, that means
that something is inconsistent in your environment and you should investigate
the cause of the inconsistency.